refactoring vectorstore loader function
